refactor: API path 경로와 같은 package 구조를 같도록 수정함
